requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
chromadb>=0.4.22
pydantic>=2.0.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
Pillow>=10.0.0
rich>=13.0.0
click>=8.1.0
pandas>=2.0.0
aiogram>=3.0.0